[Refactor] pages (#16)
* transform to styled-components

* rename Capital
