+++ src/javascripts/components/pages/users/CommonStyleComponents.ts
@@ -1,0 +1,6 @@
+import styled from "@emotion/styled";
+
+export const Root = styled.div`
+  width: 100%;
+  height: 100%;
+`;
