+++ src/stylesheets/page_container.scss
@@ -2,11 +2,6 @@ body {
   background-color: #0D171F !important;
 }
 
-#page-container {
-  width: 100%;
-  height: 100%;
-}
-
 #project-page-container {
   width: 100%;
   height: 100%;
